fix: adjust finance logics and other fixes

+++ src/finance/finance.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateBudgetDto } from './dto/budget.dto';
 import { BudgetDistribution } from './entities/budgetDistribution.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -23,6 +23,7 @@ import {
   statuses,
 } from '../users/user.interface';
 import { monthNames } from '../utility/constants';
+import { NotificationService } from '../shared/notification/notification.service';
 
 @Injectable()
 export class FinanceService {
@@ -36,6 +37,7 @@ export class FinanceService {
     @InjectRepository(DisbursementDistribution)
     private readonly disbursementDistributionRepository: Repository<DisbursementDistribution>,
     private readonly studentsService: StudentsService,
+    private readonly notificationService: NotificationService,
   ) {}
 
   async createBudget(
@@ -110,6 +112,16 @@ export class FinanceService {
 
   async deleteBudget(id: string): Promise<IResponse<Budget>> {
     const budget = await this.budgetRepository.findOneByOrFail({ id });
+    const disbursement = await this.disbursementRepository.findOneBy({
+      budget: {
+        id: id,
+      },
+    });
+    if (disbursement) {
+      throw new ConflictException({
+        message: 'Budget with disbursement cannot be deleted',
+      });
+    }
     await this.budgetRepository.remove(budget);
     return {
       message: 'Budget deleted successfully',
@@ -153,8 +165,11 @@ export class FinanceService {
   }
 
   public async getDisbursement(id: string): Promise<IResponse<Disbursement>> {
-    const disbursement = await this.disbursementRepository.findOneByOrFail({
-      id,
+    const disbursement = await this.disbursementRepository.findOneOrFail({
+      where: {
+        id,
+      },
+      relations: ['student'],
     });
 
     return {
@@ -165,11 +180,25 @@ export class FinanceService {
 
   async getDisbursements(
     page: number = 1,
+    status: string,
+    search: string,
   ): Promise<IResponse<IPagination<Disbursement[]>>> {
     const skip = (page - 1) * 10;
     const queryBuilder =
       this.disbursementRepository.createQueryBuilder('disbursement');
     queryBuilder.innerJoinAndSelect('disbursement.student', 'student');
+    queryBuilder.innerJoinAndSelect('student.school', 'school');
+    if (status) {
+      queryBuilder.andWhere('disbursement.status = :status', {
+        status,
+      });
+    }
+
+    if (search) {
+      queryBuilder.andWhere('LOWER(student.name) LIKE LOWER(:search)', {
+        search: `%${search}%`,
+      });
+    }
     const [disbursements, total] = await queryBuilder
       .skip(skip)
       .take(10)
@@ -251,6 +280,41 @@ export class FinanceService {
     };
   }
 
+  async editDisbursement(
+    id: string,
+    createDisbursementDto: CreateDisbursementDto,
+  ): Promise<IResponse<Disbursement>> {
+    const disbursement = await this.disbursementRepository.findOneByOrFail({
+      id,
+    });
+    const budget = await this.budgetRepository.findOneByOrFail({
+      status: statuses[0],
+    });
+    const student = await this.studentsService.getStudentById(
+      createDisbursementDto.studentId,
+    );
+    disbursement.amount = createDisbursementDto.amount;
+    disbursement.budget = budget;
+    disbursement.student = student;
+    disbursement.disbursementDistribution = await Promise.all(
+      createDisbursementDto.disbursementDistribution.map(
+        async (distribution) => {
+          const newDistribution = new DisbursementDistribution();
+          newDistribution.amount = distribution.amount;
+          newDistribution.title = distribution.title;
+          return this.disbursementDistributionRepository.save(newDistribution);
+        },
+      ),
+    );
+
+    await this.disbursementRepository.save(disbursement);
+
+    return {
+      message: 'You have successfully edited a disbursement',
+      data: disbursement,
+    };
+  }
+
   public async createBeneficiaryDisbursement(
     id: string,
     createBeneficiaryDisbursementDto: CreateBeneficiaryDisbursemenDto,
@@ -281,6 +345,16 @@ export class FinanceService {
     await this.budgetRepository.save(budget);
     await this.disbursementRepository.save(disbursement);
 
+    await this.disbursementRepository.save(disbursement);
+    const student = await disbursement.student;
+    const user = await this.studentsService.findUser(student.id);
+    if (user)
+      await this.notificationService.sendApproveDisbursementEmail(
+        user.email,
+        user.name,
+        disbursement.amount.toString(),
+      );
+
     return {
       message: 'Disbursement approved successfully',
       data: disbursement,
@@ -290,11 +364,18 @@ export class FinanceService {
   public async declineDisbursement(
     id: string,
   ): Promise<IResponse<Disbursement>> {
-    const disbursement = await this.disbursementRepository.findOneByOrFail({
-      id,
+    const disbursement = await this.disbursementRepository.findOneOrFail({
+      where: { id },
     });
     disbursement.status = 'declined';
     await this.disbursementRepository.save(disbursement);
+    const student = await disbursement.student;
+    const user = await this.studentsService.findUser(student.id);
+    if (user)
+      await this.notificationService.sendDeclineDisbursementEmail(
+        user.email,
+        user.name,
+      );
 
     return {
       message: 'Disbursement declined successfully',
